Download tutorials, comics and more
You may need to download tutorials for something from a site without clicking one link at a time/maybe some photos in a site with some specific kind etc. And yes, you may be a member of the site and only members may be able to download files. So this code will help you to login to such sites and download the files one by one, and yes, with status bar.